Fix issues with using Launcher3 as default launcher
If Launcher3 is set as default, it needs LOCAL_PRIVILEGED_MODULE=true to be
able to add app widgets without requiring a security confirmation from the user.

It also needs android:requiredForAllUsers="true" to force the app to be present
for restricted user profiles. Otherwise, individual activities will be listed in the
restricted profile setup screen and turned off by default. If they are not manually
enabled, then the entire launcer3 package will not be enabled and the user will have
constant crashes since there will not be a default launcher.

Change-Id: Id9cd308f57a03d656dddd0aab59cc6531deac9fa
Signed-off-by: Joel Klussman <joel.klussman@intel.com>
Signed-off-by: Dan Liang <dan.liang@intel.com>
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>